+++ resources/views/components/site/shell.blade.php
@@ -1,0 +1,13 @@
+@props([
+    'flush' => false,
+])
+
+<div
+    data-site-shell
+    {{ $attributes->class([
+        'mx-auto w-full max-w-7xl',
+        'px-4 sm:px-6 lg:px-8' => ! $flush,
+    ]) }}
+>
+    {{ $slot }}
+</div>